requirements.txt
flask
pyyaml
requests
pandas
numpy
binance-futures-connector
binance-connector
flask-socketio